factor out exports from html js
